+++ Sources/Utilities/KMCellConfiguration.swift
@@ -11,4 +11,5 @@ public enum KMCellConfiguration {
     
     /// if true then incoming messsage's sender name will be hidden on conversation
     public static var hideSenderName = false
+    
 }