+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # CHANGELOG
 
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
+## [1.0.6] 2023-04-19
+- Added Support for delete conversation for End User
+- Fixed Form text area data not added with message for `postFormDataAsMessage` issue
+- Matched form action flow with web & android.
+- Fixed submitted form data double stringified issue
+- Added Support for Static Top Message.You can enable this by using below code
+```swift
+KMConversationScreenConfiguration.staticTopMessage = "<Your Message>"
+KMConversationScreenConfiguration.staticTopIcon = "<pass your UIImage>"
+```
+## [1.0.5] - 2023-04-05
+- Upgraded Kingfisher pod to version 7.6.2
 
 ## [1.0.5] - 2023-04-05
 - Upgraded Kingfisher pod to version 7.6.2